move where patching code for xpu in detection
