+++ src/otx/algorithms/detection/adapters/mmdet/apis/train.py
@@ -7,8 +7,6 @@
 import os
 
 import torch
-from mmcv.ops.nms import NMSop
-from mmcv.ops.roi_align import RoIAlign
 from mmcv.runner import (
     DistSamplerSeedHook,
     EpochBasedRunner,
@@ -16,18 +14,14 @@ from mmcv.runner import (
     build_runner,
     get_dist_info,
 )
-from mmcv.utils import ext_loader
 from mmdet.core import DistEvalHook, EvalHook, build_optimizer
 from mmdet.datasets import build_dataloader, build_dataset, replace_ImageToTensor
 from mmdet.utils import build_ddp, compat_cfg, find_latest_checkpoint, get_root_logger
 from mmdet.utils.util_distribution import build_dp, dp_factory
-from torchvision.ops import nms as tv_nms
-from torchvision.ops import roi_align as tv_roi_align
 
 from otx.algorithms.common.adapters.mmcv.utils import HPUDataParallel, XPUDataParallel
 from otx.algorithms.common.adapters.mmcv.utils.hpu_optimizers import HABANA_OPTIMIZERS
 
-ext_module = ext_loader.load_ext("_ext", ["nms", "softnms", "nms_match", "nms_rotated", "nms_quadri"])
 dp_factory["xpu"] = XPUDataParallel
 dp_factory["hpu"] = HPUDataParallel
 
@@ -134,11 +128,6 @@ def train_detector(model, dataset, cfg, distributed=False, validate=False, times
     # build optimizer
     auto_scale_lr(cfg, distributed, logger)
 
-    if cfg.device in ["hpu", "xpu"]:
-        # dynamic patch for nms and roi_align
-        NMSop.forward = monkey_patched_nms
-        RoIAlign.forward = monkey_patched_roi_align
-
     optimizer = build_optimizer(model, cfg.optimizer)
 
     if cfg.device == "xpu":
@@ -211,50 +200,3 @@ def train_detector(model, dataset, cfg, distributed=False, validate=False, times
     elif cfg.load_from:
         runner.load_checkpoint(cfg.load_from)
     runner.run(data_loaders, cfg.workflow)
-
-
-def monkey_patched_nms(ctx, bboxes, scores, iou_threshold, offset, score_threshold, max_num):
-    """Runs MMCVs NMS with torchvision.nms, or forces NMS from MMCV to run on CPU."""
-    is_filtering_by_score = score_threshold > 0
-    if is_filtering_by_score:
-        valid_mask = scores > score_threshold
-        bboxes, scores = bboxes[valid_mask], scores[valid_mask]
-        valid_inds = torch.nonzero(valid_mask, as_tuple=False).squeeze(dim=1)
-
-    if bboxes.dtype == torch.bfloat16:
-        bboxes = bboxes.to(torch.float32)
-    if scores.dtype == torch.bfloat16:
-        scores = scores.to(torch.float32)
-
-    if offset == 0:
-        inds = tv_nms(bboxes, scores, float(iou_threshold))
-    else:
-        device = bboxes.device
-        bboxes = bboxes.to("cpu")
-        scores = scores.to("cpu")
-        inds = ext_module.nms(bboxes, scores, iou_threshold=float(iou_threshold), offset=offset)
-        bboxes = bboxes.to(device)
-        scores = scores.to(device)
-
-    if max_num > 0:
-        inds = inds[:max_num]
-    if is_filtering_by_score:
-        inds = valid_inds[inds]
-    return inds
-
-
-def monkey_patched_roi_align(self, input, rois):
-    """Replaces MMCVs roi align with the one from torchvision.
-
-    Args:
-        self: patched instance
-        input: NCHW images
-        rois: Bx5 boxes. First column is the index into N. The other 4 columns are xyxy.
-    """
-
-    if "aligned" in tv_roi_align.__code__.co_varnames:
-        return tv_roi_align(input, rois, self.output_size, self.spatial_scale, self.sampling_ratio, self.aligned)
-    else:
-        if self.aligned:
-            rois -= rois.new_tensor([0.0] + [0.5 / self.spatial_scale] * 4)
-        return tv_roi_align(input, rois, self.output_size, self.spatial_scale, self.sampling_ratio)
