+++ src/otx/algorithms/detection/adapters/mmdet/utils/__init__.py
@@ -11,6 +11,8 @@ from .config_utils import (
     patch_ir_scale_factor,
     patch_tiling,
     should_cluster_anchors,
+    monkey_patched_nms,
+    monkey_patched_roi_align,
 )
 
 __all__ = [
@@ -21,4 +23,6 @@ __all__ = [
     "patch_input_shape",
     "patch_ir_scale_factor",
     "should_cluster_anchors",
+    "monkey_patched_nms",
+    "monkey_patched_roi_align",
 ]
